[WIP] First pass implementation for RAS v1.1
This basically works for fetching the v1.1 tokens, but I want to do
another pass and make sure a few more things work:

* I need to double check the RAS Passport is being validated correctly
with the correct fields
* The newest implementation isn't backwards compatible with v1.0, I
want to fix that.
* I might change the exception handling a bit, I think auth should fail
if passport information doesn't verify but I want to clean that up a bit.
* I noticed some funny business when requesting the invalid scope
'passport_jwt_v11', where userinfo returned an invalid response. That
may be worth submitting a ticket to RAS if it wasn't a simple mistake
on my end somewhere.